+++ Samples~/Sounds/Music/Scripts/MUSICExtensions.cs
@@ -6,12 +6,7 @@ namespace Audio.Music
     {
         public static void PlayMusic(this MUSIC music)
         {
-            var musicController = MusicController;
-
-            Assert.IsNotNull(musicController, $"Missing the {nameof(MusicController)} in the Scene!!");
-            musicController.PlayMusic(music);
+            MusicController.PlayMusic(music);
         }
-
-        private static MusicController MusicController => MusicController.Instance;
     }
 }