+++ Samples~/Sounds/SoundFx/Scripts/SFXManager.cs
@@ -6,12 +6,15 @@ using Sounds;
 using UnityEngine;
 using UnityEngine.Assertions;
 using UnityEngine.Audio;
+using Utilities;
 using Random = UnityEngine.Random;
 
 namespace Audio
 {
-    public class SFXManager : MonoBehaviour, ISetVolume
+    public class SFXManager : HiddenSingleton<SFXManager>, ISetVolume
     {
+        internal static SFXManager instance => Instance;
+        
         //============================================================================================================//
         [Serializable]
         private class SfxData
@@ -43,7 +46,6 @@ namespace Audio
         [SerializeField]
         private AudioSource sfxSourcePrefab;
         private List<AudioSource> _audioSources;
-        internal static SFXManager Instance;
 
         [SerializeField] private SfxData[] sfxDatas;
 
@@ -55,22 +57,12 @@ namespace Audio
         //Unity Functions
         //============================================================================================================//
 
-        private void Awake()
-        {
-            if (Instance != null)
-            {
-                Destroy(gameObject);
-                return;
-            }
-
-            Instance = this;
-            Assert.IsNotNull(sfxAudioMixer);
-            
-        }
-
         // Start is called before the first frame update
         private void Start()
         {
+            Assert.IsNotNull(sfxAudioMixer);
+
+            
             InitVfxLibrary();
 
             _sfxAntiSpam = new Dictionary<SFX, int>();
@@ -94,8 +86,12 @@ namespace Audio
         }
 
         //============================================================================================================//
-
-        internal void PlaySound(SFX sfx, float volume = 1f)
+        public static void PlaySound(SFX sfx, float volume = 1f)
+        {
+            Assert.IsNotNull(Instance, $"Missing the {nameof(SFXManager)} in the Scene!!");
+            Instance._PlaySound(sfx, volume);
+        }
+        private void _PlaySound(SFX sfx, float volume)
         {
             if(!_isReady)
                 return;
@@ -119,8 +115,13 @@ namespace Audio
             //FIXME This should just use a loop, and not a coroutine
             StartCoroutine(DequeueSFXCoroutine(sfx, audioClip.length));
         }
+        public static void PlaySoundAtLocation(SFX vfx, Vector3 worldPosition)
+        {
+            Assert.IsNotNull(Instance, $"Missing the {nameof(SFXManager)} in the Scene!!");
+            Instance._PlaySoundAtLocation(vfx, worldPosition);
+        }
         //This is meant to be called via the VFXExtensions class
-        internal void PlaySoundAtLocation(SFX vfx, Vector3 worldPosition)
+        private void _PlaySoundAtLocation(SFX vfx, Vector3 worldPosition)
         {
             if(!_isReady)
                 return;